refactor: Reorganized components in the folder structure
